logic separated from DOM manipulation
